requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
psycopg[binary,pool]==3.2.3
python-dotenv==1.0.1
openai==1.75.0